+++ CHANGELOG.md
@@ -1,3 +1,24 @@
-# CHANGELOG
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+- Test 8
+- Test 7
+- Test 6
+
+## [0.0.2] - 2024-08-12
+
+- Test 5
+- Test 4
+
+## [0.0.1] - 2024-08-12
+
+- Test 3
+- Test 2
+- Test 1